BaseController: total refactor
Make it a class instantiated with a Registry, and have it use the
registry to find an adapter (and urlTemplates etc) required to handle
each request. No more janky extend/subclasses/adapterFn on the
controller.

Add basic support for create and delete. (Still waiting on side
posting, and delete doesn’t manage associations or call userland code
that could do so)