MongooseAdapter: total refactor
Now that it’s clear that this is going to have to handle much more than
GETs, have top level methods only for crud operations (and some static
helpers), and keep queryBuilder state in the find method.

Support the concept of extraResources—ones included in the response but
not directly connected to any primary resources.

Start create and delete support