MongooseAdapter: don't add a subType property after all
Adding subType was stupid: it opened the door to possible naming
conflicts, it arguably exposed implementation details that clients
shouldn’t know about, and it was a strategy that might have been
incompatible with other inheritance models/adapters. The only reason I
introduced it in the first place was to help with issue #30, but the
prior commit’s `processAsType` property handles that way more elegantly.