Finally, a solution to child type handling
The processAsType hack didn’t end up working because, while it allowed
a `SubTypeOfA` to be processed as such even while it was being returned
as `TypeOfA` in response to a request for `A`s, it didn’t allow
`SubTypeOfA`’s returned in response to requests for `SubTypeOfA`s to
know that they were also `A`s. I.e. it didn’t expose the parent type,
so a subtype couldn’t also be processed as its parent type before
return.

Now, we actually have Resource.type represent the item’s real type (the
deepest node it belongs to on the type tree) and the
ResourceTypeRegistry stores each type’s parentType to allow for
recursive processing. This was easier than trying to get the parentType
from the adapter, which can’t always (quickly) deduce it.