Clarify component roles
Introduce a Pipeline who’s job is to take a RequestContext and produce
a ResponseContext that has everything needed to generate the response.
Introduce a controller that’s responsible for generating the
RequestContext from the raw request object and the response from the
ResponseContext object, that can act as an interface with Express. And
move the req -> RequestContext object logic out of RequestContext and
into this new controller.