Major Controllers/Pipeline refactor
Before, there was the concept of an APIController that interfaced with
the framework and then passed framework-neutral objects to a Pipeline
for handling. The problems with this were:

1. We also had a Documentation controller that needed the same
shielding from the framework that the APIController was giving the
pipeline code, but I was loathe to repeat APIController code in the
Documentation controller or merge the Documentation handling code into
a controller called APIController.

2. The Pipeline, while a reasonable abstraction and a reasonably simple
one to wrap one’s head around, was a non-standard abstraction, and if
non-standard abstractions can be avoided, it makes the library easier
to contribute to.

So now, we have a Front controller, which is constructed with instances
of the “child controllers” API and Documentation. Its job is to handle
all requests from the framework by converting the framework’s
request/response objects to neutral ones, calling `handle` on either
the Documentation controller or the API controller (which act as what
used to be called Pipelines), and turning those child controllers’
Response objects into real framework responses.

This mirrors, for example, the pattern Symfony uses in the PHP world to
take all of PHP’s $_GET/$_POST nonsense and turn it into real
request/response objects that each controller can work with.

Finally, I renamed RequestContext and ResponseContext to simply Request
and Response, which is a little disingenuous since they don’t have any
behavior (i.e. methods), but is more in keeping with the other types’
names. Also, down the line, we might choose to add the behavior (like
send, set, etc) that’s currently—and somewhat inappropriately, because
it’s simpler pragmatically—being handled in the FrontController.