API Controller: Implement New Content Negotiation Logic
1. Pick response Content Type at the start of the request, rather than
at the end. We were picking it at the end before on the theory that an
extension could be turned on in the process of handling the request,
and that would change some media type param on the Content-Type. But
that possibility’s gone for now, with the the whole extension system.

2. The only exception to (1) is that, if we have an error, we can only
produce that error in JSON (API), even if the endpoint could’ve
produced/negotiated for something else (e.g. the Documentation
controller, which isn’t actually using this logic yet but should be,
could’ve negotiated for html). So we always set the error content-type
to json or json api explicitly at the end.