Tests: circumvent old v8 bug
V8 used to delete an object entirely sometimes when its last key had
been deleted (https://code.google.com/p/v8/issues/detail?id=2980), as
we were doing in this test. And this issue still existed on pretty
recent versions of node, so this commit just rewrites the test case to
not trigger the bug.