Tests: fix loading (it's the order, stupid)
Using before() ensures that the fixtures are loaded before each test
(i.e. each `it()`) runs. But it doesn’t ensure that the fixtures are
loaded before each use of the Agent happens in the test files. And,
because the agent is used outside the `it()` calls, so we can make one
request and assert many things about it, we have to make sure the
fixtures are loaded before we even require those other files that use
the agent. That’s what this change does. Note that we can’t use
`import`, but need require instead, because `import` needs to be at
top-level.