Error type [Breaking]: reveal details less liberally
Previously, we only checked `isJSONAPIReady` before reading in the
error’s `message`—and even then only for the `title`, not for
`details`—but we didn’t require `isJSONAPIReady` to migrate over the
`code` and other properties. This was a security risk, since APIErrors
are displayed to the client directly, so they should only expose info
that the user has explicitly signaled is safe.

This commit also stops referencing err.message for both the JSON API
Error’s `title` and `details` property; instead, `message` can only go
into `details`. This is more consistent with JSON API semantics and
prevents the odd case in which `title` and `details` are the same