ApiError.fromError: don't pass through blank messages
If a generic `Error` object is created with no message, e.g. `new
Error()`, the `message` property is internally set to an empty string.
So, in the conversion, we were getting the `detail` property
automatically set to an empty string. Now, the detail property is left
blank and so not displayed because we make sure message is truthy
before using it.