BREAKING: APIController.responseFromExternalError signature change
Previously, responseFromExternal error required a framework-neutral
Request object as an argument. But that was circular if the
error-to-be-reported occurred _while constructing the framework-neutral
request object_ from the incoming http message, as can happen in
ExpressStrategy when body parsing fails. So this now changes the
function to only require the accept header, which is all the info it
actually needed (no law of demeter violation) and which can always be
produced without causing an error.