Express Strategy: more rigorous empty body detection
Earlier, when we changed `hasBody` to `hasNonEmptyBody` body, we were
identifying empty bodies just by checking if Content-Length === 0. But
that’s insufficient because, when Transfer-Encoding: chunked is being
used, there won’t be a Content-Length, and yet an empty body can still
be sent. So, this commit reverts `hasNonEmptyBody` to `hasBody` and
then does the emptiness check once getRawBody is done reading the
(possibly transfer-encoded) stream.

Also, if the body doesn’t exist, this now sets the request’s `.body`
property to undefined instead of null, since null is actually a valid
json value that (theoretically) could’ve been parsed from the response.
Also, if the body is present but empty, this now sets the `.body`
property to the empty string, whereas before it was also set to `null`,
in case the code needs to differentiate between an empty body and a
non-existent one.

Also, note that this means that now: `request.hasBody === (request.body
=== "" || typeof request.body === "undefined")`, whereas before the
equality was `request.hasBody === (request.body === null))`