[Breaking]: Support alternate relationship forms
A PR for #97. This:

1. Renames RelationshipObject to Relationship (the old name was clunky
for now reason).

2. Exports Relationship and Linkage for the first time, such that
they’re now part of the library’s public API.

3. Allows Relationships to specify uri templates that override the
resource level ones (and renames the Relationship class fields
accordingly).

4. Allows relationships to be output without data (if linkage is
undefined on the Relationship).

This is only breaking in the sense that user-defined adapters or
beforeSave/beforeRender methods might have been depending on
Relationships to have their old field/class name.