ResourceTypeRegistry/pseudoTopSort: less radical reordering
Sometimes, the order in which the resource type descriptions are given
to the registry—which is also the order they’re listed in in the
documentation—is significant to the user, so we try to preserve it as
much as possible. This may be true in general for graphs sorted with
pseudo top sort, and the performance cost is small, so we just make the
change there.