Support Query.resultsIn and split Response into Result & HTTPResponse
This is a big improvement. Here’s the context:

Earlier in v3, we introduced "query transforms". Basically, the idea is
that the library has always inspected the request and implicitly
constructed a query from it (in the form of arguments passed to an
adapter method), but users had no way to customize this query. That
proved to be a big limitation. One obvious example: suppose you want to
filter the results for a `GET /some-collection` request to only include
the entries that a user has permission to see; you could do this in
`beforeRender` but: 1) this was very slow, because you were still
fetching all the items from the db and 2) I believe that messes up the
pagination count. With query transforms, you can just transform the
library's generated query to have an extra where clause, so that only
the should-be-visible collections are fetched.

Even with this system in place, though, the user can't instruct the
library to perform multiple queries at once (e.g. in order to update
the inverse side of a relationship when one side is updated). I plan to
address that soon, and this commit lays the groundwork for that. In
particular, once the user can perform multiple queries, it becomes
obvious that there needs to be someway for the user to customize where
the results of each query end up in the response document. (E.g., the
result of the query to update the inverse side of the relationship can
probably be discarded, whereas the results of another query might need
to be added to `meta` as supplementary data.) To support that, this
commit stores two functions on all query objects — returning and catch
— that are called with the query’s results, on success and failure
respectively, and get to return an object indicating where the query
results/error should go in the response. These properties are given a
default function for all queries generated by the library, but they can
be replaced in a query transform, like any other query property, using
the `Query.resultsIn` function to set one or both of them immutably.
See changelog for how to do this.

One cool thing about this is that it solves #136, in that the user can
register a function to transform any query errors as needed.

Note: these returning/catch functions must return an instance of the
newly-defined Result type (see types/index.ts).

Note 2: I had to call the function that runs on a successful query
`returning` rather than `then()`, which would’ve been my name of
choice, because various JS syntax constructs (e.g. `await`) treat
objects with a `then()` method specially, in ways we don’t want, due to
historical stuff around how Promises/“thenables” were used before
Promises were added to ES6.