APIError: cleanup + extend validation logic; support simpler construction syntax
Now that even Node 6 supports Proxy, we can use it safely without
having to do those crazy hacks to validate input while preserving the
enumerability + own property setup that we want.

We also support an object in the constructor because, with so many
optional properties, that’s just more convenient.