types: Add ResourceIdentifierSet and support it as Document.primary
When we construct a document for a bulk delete request, the primary
data is ResourceIdentifier objects, and yet it’s not a Relationship
(there’s no owner), so we need a type to represent that.
ResourceIdentifierSet is that type.

Also, validate document.meta, because we may be constructing documents
from JSON soon, if we're gonna but the whole request body into a
Document (which is why the ResourceIdentifierSet is needed.)