fix: Query building resultsIn bug
The query.returning and query.catch functions on the queries that our
makeXXX() functions were returning happened to always be synchronous,
but the type signature allows them to be async. So, to make sure things
don’t blow up if we make one of them async in the future, use proper
async function composition to build the final .returning/.catch.