refactor: tweak model name <=> type name mapping logic
BREAKING CHANGE.

Throughout the adapter, we need to map the `type` value the adapter
functions receive (e.g., in Query objects) to mongoose models. This
requires knowing the model name for each type. (We can’t just have the
user provide a { typeName: model } mapping because mongoose’s
discriminator keys always store the model name in the db.)

A fair question, then, is: should this concept of a model name be
purely local to the adapter (with the possibility, potentially, that
some adapters wouldn’t even have/need this concept), or should the
library writ large know that each type name also has a model name? The
former is clearly simpler, but, at present, the model name is used
outside the adapter to generate documentation and, in that context,
having a name for the type that’s singular and pascal cased is kinda
nice.

So, let’s say we keep the concept of a model name for each type; then
the question becomes: where should that model name live?

1. It could continue to be generated by the adapter, with a function on
that interface.

2. It could live on the resource description. This has the big
advantage of not requiring adapters that don’t have the concept of a
model name to reimplement (and allow the reconfiguration of) inflector
transformations. It has the downside of requiring the user to specify a
model name manually in each type description, but that can easily be
mitigated by instead setting a `getModelName(typeName)` function on
type descriptions (rather than a model name string). That function can
then be given a default that’s automatically inherited in all the
resource type descriptions. Bam!

The question with option 2 is: how does the adapter get the modelName
=> typeName mapping? The user could pretty easily construct such a
mapping from the resource descriptions/methods on the resource type
registry, but I don’t want the user to have to do that manually (nor do
I want the registry to be an argument to the adapter). So, I want to
somehow have a { modelName: typeName } mapping that can be assigned a
default, automatically-computed value that works in most cases, so the
user doesn’t have to do anything. The way to do that, I think, is to
have the MongooseAdapter take as its second argument a function that is
`toTypeName(modelName)`. That can then be defaulted to a function
that’s the inverse of the default `toModelName` function we use in the
registry.

This has the advantage too of letting the modelName used in the
registry (for documentation) not be the same as the mongoose model name
— although can’t imagine that’s very useful in practice.

However, by not having the map argument derived from the registry
methods directly, if the user changes the registry’s `toModelName`
logic (not to have the Mongoose model name differ from the model name
for documentation purposes but, e.g., to fix an inflector bug), they
also have to pass a different `toTypeName` function to the adapter. I
*think* that’s worth it for the ability to default the adapter mapping.

A third solution, which is very similar, would be have the docs
controller to take a `toModelName` function that works and would be
defaulted the same way as the resource type registry’s corresponding
function as proposed above. This likewise enables the model name in the
adapter and the docs controller to both be defaulted and to vary
independently, at the cost that an inflection rule change may have to
be made in both places. The only real difference between it and option
2 is whether we want to have a registry method for getting model
names/conceptually whether we want model name to be a type description
property vs a documentation-only thing. I think documentation only is
more appropriate.

So, I’ve implemented solution 3 here. This introduced a few breaking
changes; see Upgrading.md.