fix(typings): undo typings tweaks
We actually can continue to use `this` for convenience. I thought/think
it might not be the right type, strictly speaking, because we're
creating a new instance -- but maybe typescript assumes that a `this`
return type is just structurally the same, rather than actually the same
reference. Either way, it's worth it, because not using this creates
more type errors unless we redefine every method in every subclass.