feat: support result factories
Until now, the structure has been:

1. The library parses the request (parsing + validating the json,
parsing query parameters, etc). There’s no documented way to customize
this behavior, with the exception of the filter parameter parsing part,
though one could subclass the API controller.

2. That “finalized request” is turned into a single query. This
behavior can be customized per request, by passing a “query factory”
function.

3. The generated query is run, its returns a Result (through
query.returning) and that Result is sent to the user.

The problem with this model was that custom “off-spec” requests (e.g.,
for sign-in flows or whatever) often need to perform queries that can’t
currently be generated by a query factory. For example, a custom
request might need to make two queries in parallel, error if either
matches no records, and merge the results if both return successfully.
That can’t be expressed in our current query semantics.

So, this commit let’s the user provide a function for generating the
Result, in the same way query factory functions let them provide a
function to build a query.

These two extension compose well — the default result factory (which
holds the library’s previously-hardcoded logic) uses the user’s custom
query factory, or falls back to the built-in one; likewise, the user
can provide a custom query factory and result factory, and the custom
result factory is passed the custom query factory.

One reasonable question: if the user’s providing a function for making
the result, why should they send the request through the library at
all? And the answer is that the library is still: 1) doing the request
parsing; 2) providing helper functions to the user (as a request
factory param) for running transforms etc; 3) formatting the Document
the user constructs as part of their result into JSON:API JSON.

Note: there is ExpressStrategy.sendResult, which lets the user respond
with a custom result, but doesn’t given them any access to key
functions for running transforms, using the parsed request, etc. Given
that, I may remove ExpressStrategy.sendResult altogether to keep the
surface area down.