refactor: new query transform interface
BREAKING CHANGE. This commit removes the long-deprecated
`transformedAPIRequest` method on ExpressStrategy, but adds a
queryTransform field to the options object passed to
`customAPIRequest`.

The API controller just desugars any provided query transform to a
query factory that first calls the library’s built-in `makeQuery`
function. This simple desugaring is why query transforms were
deprecated to begin with; however, my experience has been that wanting
to just transform the library’s default query is common enough that
it’s worth having a interface for it that removes the query factory
boilerplate.