refactor: Remove parsing of incoming links
This parsing was: 1) wrong and therefore would’ve been insecure if we
were actually using the parsed values for anything (it assumed the link
was always a string, but it could be a link object or null etc); 2)
useless, as I can’t imagine a scenario where it would make sense for
the client to provide links (atm) and the server to respond to them. If
we find such a scenario later, we can add this back properly.