refactor: url templates
BREAKING CHANGE. Now, the registry accepts templates as strings (like
before) or functions (new), but always stores them as a new UrlTemplate
type, which encodes both the parsed template function and the original
string (if any) for showing to the client/in documentation. The rest of
our code, which reads the templates (directly or indirectly) from the
Registry, now always work with the function version. Most code already
expected a function version (since we did the conversion to functions
pretty early in the API controller), but now the last little bits do as
well (e.g., make-post.ts).

Types & fields have been renamed/removed to capture this new setup.