feat(request-parsing): support user providing request supported operators
While it’s true that the query produced by a request can’t include any
operators that are unsupported by the adapter that runs the query, that
doesn’t mean that the request itself can’t have unsupported operators
in its url. (I.e., the request may contain an “unsupported” operator as
long as that operator doesn’t make it into the query that’s ultimately
run.) To support this case, and to recognize that request parsing
shouldn’t be inherently coupled to adapter capabilities, we rework
things so that the user can explicitly provide a set of supported
operators for each request, with the adapter’s supported operators just
acting as a default set.

This also kills the cache for finalizing the operators, because that
really seems like premature optimization (and not even a guaranteed
speed win), even though we are re-finalizing the operators now on every
request.